ml/requirements.txt
pydicom>=2.4.0
SimpleITK>=2.3.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0
pynrrd>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
matplotlib>=3.7.0
